Add status param support
This commit adds "status" param support. It ditches the current behavior,
that when time is set, the device already starts working. Instead, the
temperature and time can be now set independently and dailing them in is
postponed untill status start is set. The status stop can be used to
stop the device. If the device is running allready then changes to time
and teperature are done live.

Signed-off-by: Mateusz Kusiak (Timax) <timax2@wp.pl>